move structures used for fuzz to common crate
Created a new crate for all the structure defines that we need during
fuzzing. This change is required so that we can use the structures
for generating custom input. Functions & structures defined in the
fuzz_targets directory cannot be exported. Also, running any other
function besides the main libfuzzer function from the fuzz_targets
directory is not permitted. Thus to be able to write custom fuzz
tests to a file on request, we need to move the definitions to
a place where they can both run on demand and be imported by the
fuzz targets.
