+++ fuzz/common/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "common"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+bincode = "1.3.3"
+serde = "1.0.63"
+virtio-queue = { path = "../../crates/virtio-queue", features = ["test-utils"] }
+virtio-vsock = { path = "../../crates/devices/virtio-vsock" }
+virtio-queue-ser = { path = "../../crates/virtio-queue-ser" }
+vm-memory = { version = "0.9.0", features = ["backend-mmap", "backend-atomic"] }
