+++ fuzz/common/src/virtio_queue.rs
@@ -1,0 +1,9 @@
+use crate::{FuzzingDescriptor, VirtioQueueFunction};
+use serde::{Deserialize, Serialize};
+
+/// Input generated by the fuzzer for fuzzing virtio_queue
+#[derive(Serialize, Deserialize, Debug)]
+pub struct VirtioQueueInput {
+    pub descriptors: Vec<FuzzingDescriptor>,
+    pub functions: Vec<VirtioQueueFunction>,
+}
