+++ fuzz/fuzz_targets/virtio_queue.rs
@@ -1,16 +1,8 @@
 #![no_main]
+use common::virtio_queue::VirtioQueueInput;
 use libfuzzer_sys::fuzz_target;
-use rust_vmm_fuzz::{FuzzingDescriptor, VirtioQueueFunction};
-use serde::{Deserialize, Serialize};
 use virtio_queue::{mock::MockSplitQueue, Descriptor};
 use vm_memory::{GuestAddress, GuestMemoryMmap};
-
-/// Input generated by the fuzzer for fuzzing virtio_queue
-#[derive(Serialize, Deserialize, Debug)]
-pub struct VirtioQueueInput {
-    pub descriptors: Vec<FuzzingDescriptor>,
-    pub functions: Vec<VirtioQueueFunction>,
-}
 
 fuzz_target!(|data: &[u8]| {
     let fuzz_input = match bincode::deserialize::<VirtioQueueInput>(data) {
